pyproject.toml
[tool.poetry]
name = "x-tender-proposal"
version = "0.1.0"
description = ""
authors = ["v20200315 <v20200315@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.9"
langgraph = "^0.2.53"
langchainhub = "^0.1.21"
langchain-community = "^0.3.9"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.4"
langchain-openai = "^0.2.10"
pypdf = "^5.1.0"
langgraph-checkpoint-sqlite = "^2.0.1"
streamlit = "^1.40.2"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
